Added a shim for CurrentBatchIterator
